move `Legacy` structs to `schema_change.rs`
